I added the priority class and added it to the Bind in the ShoppingListItem Controller.
