+++ ShoppingList/ShoppingList/Models/ShoppingListItem.cs
@@ -14,16 +14,8 @@ namespace ShoppingList.Models
         [MaxLength(8000)]
         public string Content { get; set; }
 
-        public enum Priority
-        {
-            [Display(Name = "It Can Wait.")]
-            ItCanWait = 0,
-            [Display(Name = "Need It Soon")]
-            NeedItSoon = 1,
-            [Display(Name = "Get It Now!")]
-            GetItNow = 2
-        }
-
+        public Priority Priority { get; set; }
+        
         [MaxLength(8000)]
         public string Note { get; set; }
